Cleanup path handling
Saw a lot of tmpbeta-20191122150041-ffae8e101b0737d21e83bffb256366788620ac24 folders, leading to believe that we don't always put a slash behind the tmp folder
All whcore paths are now consistent and end with a slash
A lot of paths are used in only one or two places, so we don't need all the helpers and local variables. Cleaned up a bit
Removed substitute module support, we won't be using them again but just outright archive existing modules if we have to. (Substitutes just postpone this until the substitute is cleaned up so we keep getting old versions popping up. So they didn't really solve their problem of a 'smooth' migration without having to delete modules)
Eliminated a lot of startup options, we rarely use them and they have mostly moved to envvariables anyway