Cleanup buildprocess
Never completed 'prevent-deploy-old-version' and gitlab is working on fixes in their CI process, so it'll probably never be needed. Remove it
Remove docker=1 from buildinfo, consistently use WEBHARE_IN_DOCKER envvar (I think we should completely replace the buildinfo file with an envvars for simplicity? It may even allow us to remove branding.h if done properly)
Get rid of buildertag, we're not using it and it unncessarily leaks info on builder configuration