Do not database-disable inapplicable tasks, rename posgresql task
postgresql cleanup task was never enabled on upgraded webhares because inapplicable tasks explicitly got disabled in the database. change this to not touch the enabled flag but still honor inapplicable, so the task can enable when webhare restarts.
also change the currnent postgresql task tag so it gets reeanbled everywhere