Move Tollium appinfo to separate namespace and support its validation
We're not validating the callbacksignatures because it's not sufficient to <xs:import/> the tollium XSDs for that. Those only work for the element etc. references in the XSD (ie, in the attributes: type="tc:Handler" but not for the elements themselves (<callbacksignature/>)
So we need a XSD wrapping both the xmlschema XSD and the tollium XSD, to validate the XSD that will be validating your XML. Or to be correct, that XSD now also becomes part of another XSD which wraps all component XSDs, (except for the siteprofile XSD), which we store at the fake location x-webhare-builtin-xsd:components.xsd?type=tolliumcomponents&amp;skipns=http://www.webhare.net/xmlns/publisher/siteprofile, which can then finally be used to validate your screen.
It does mean that we need to separate XSDs meant for components from other XSDs, as only the former XSDs should actually go through this tolliumcomponents.xsd.xsd.