Add option to allow checking disabled fields
By default, disabled fields are not considered when evaluating form conditions. The `checkdisabled` option can be used in conditions when a disabled fields is used as the condition source.
This can be used for example when a city field within an address is a condition source, but can only be updated indirectly by setting the zip and number fields.