Don’t use composite aggregation for terms
If it’s needed (when there are more terms than the max number of buckets), it can always be brought back, but then it can be added properly by retrieving all terms using the composite aggregation’s scroll-like feature (which wasn’t implemented anyway). By using the normal terms aggregation, it can be used with a nested terms aggregation.