Remove unused code
.. to resist the temptation to keep trying. I attempted to use it for invoking PopPrimaryFinishHandlers a bit faster and avoiding a few async callbacks. It's probably never going to be useful, if you try to do this from EM_Syscall you risk EM_Syscall returning its promise-id just after you did OpenFunctionCall, breaking the stack or VM or something else important.

and even this use case had an even faster solution - just pass a flag to CommitWork that there are handlers to request. Or transform CommitWork to an explicit external function instead of a syscall and we can pass the handler record straightaway.

So let's just put the final nail in the coffin here and eliminate this