Move map component to TypeScript, add Leaflet
This migrates the map component to TypeScript and the new Tollium iframe api. It also adds support for using OpenStreetMaps in the backend through Leaflet and makes that the default provider, so no api keys are necessary to use the map component in Tollium.

The old Google maps images, which are no longer used are removed as well.