[backport] enable avif/webp in image editor
new: allows opening AVIF (and WebP) in image editor. It will still be saved as JPEG

(no browser supports saving as AVIF yet. most support saving as WEBP. But as long as HareScript can't efficiently scandata on webp (it'll send it off to callJS) we'll just use JPEG as the storage format